Moved optional anonymization tag values to config. Only admin can export anon keys.
