+++ src/main/resources/application.conf
@@ -91,6 +91,16 @@ slicebox {
 
   anonymization {
     purge-empty-keys = false
+
+    key-attributes: [
+      { tag-path: "(0008,0050)", level: "STUDY" }, // AccessionNumber
+      { tag-path: "(0008,1030)", level: "STUDY" }, // StudyDescription
+      { tag-path: "(0008,103E)", level: "SERIES" }, // SeriesDescription
+      { tag-path: "(0010,0030)", level: "PATIENT" }, // PatientBirthDate
+      { tag-path: "(0018,1030)", level: "SERIES" }, // ProtocolName
+      { tag-path: "(0020,0010)", level: "STUDY" }, // StudyID
+      { tag-path: "(0020,0052)", level: "STUDY" } // FrameOfReferenceUID
+    ]
   }
 
   superuser {
